DW1 ver. 0.1.6 Module Updates
Change to a shorter name to avoid being the same
as the original game, and appear neater on
smaller screens.

Edit the module description.

Edit the module logo with aspect ratio and size
corrections.

Change the module flag to distinguish it more
from the DMC module, and refer to the original
game's platform.