Game Crash Status Page 4 Eggs
On many eggs, the game crashes on the Status
Window, page 4, with the error:

Game encountered an error: integer modulo by zero

This is due to hunger_loss or poop_timer being
zero.

Fixes the formulas with "% max(1,(expression))"

Changes the hunger_loss and poop_timer on
various eggs to default non-zero value.