Strength "in-need" change
Change the Feeding Menu such that "in-need"
selection will not show pets in need of vitamins
if the module's care protein_strengh_gain == 0.

Will not affect any existing modules, as the
default value for modules is 1 and all included
modules have a value of 1.

Used for my DW1 module Work-in-Progress, where
vitamins are used like "MP floppies", and
strength is only from training.