Missing Attack Animations Bugfix
For DMC and DMX, Stage 1 and Stage 2 pets without
a set attack animation have been given atk_main
"1", a "bubble" attack. This animation might not
match keychain pets, so it may need correction,
but this bugfix is simply to allow them to train.

Gives Amphimon in PENC attack animations. These
animations might not match the keychain, so they
might need correction, but this will allow
Amphimon to train and battle.