Experimental Default Weight Implementation
Modifies both the game and Module Editor to have
default weight for a pet set on evolution. Adds
the value to documentation charts if set for a
pet.

Disabled by default, and does not crash the game
if monster.json does not have the value, as it
sets it to 0, or disabled. Module Editor defaults
to 0, or disabled, when creating a new pet.

Minimum value is 0, disabled, and maximum is 99.