Pet Evolution Weight Implementation
This adds a value to Module Editor and the game
of "Evolution Weight", an optional value separate
from minimum weight, that a pet's weight is set
to when it evolves.

Add a value of 0, meaning disabled, to the pet
templates.

Evolution weight is added to the data copied
when copying a pet with Module Editor.

Evolution weight is added to the documentation
charts when it is set higher than the minimum
weight.