Evolution Editor Zero "Null" Distinction
Without this change, Module Editor erases any
manual 0,0 criteria entry in monster.json.

This changes Evolution Editor to use -1 as a
"null" value, and the default value for the
number ranges in the Edit Evolution Criteria
dialogue. When either value of a range is -1, it
returns null.

This change allows specific selection of a
criteria of 0,0, allowing that criteria to be
loaded from and saved to monster.json.

Does not save any -1 values to monster.json.