EvolutionCriteriaForm Improvements
Fixes bug that errors out Module Editor if a
number higher than 10 is entered for level, or a
number higher than 100 for WinRatio, Weight, or
QuestsCompleted.

Lowers either number of a range to the max value
if set higher than the max value.

Reduces the lower number of a range to match the
uppper number of a range when the lower number is
larger.

Changes so a criteria is disabled when the upper
number of the range is set to -1.

Automatically changes a lower number of a range
from -1 to 0 when the upper number is set >= 0.