Protein Overdose Penalty Implementation
This add the variable of protein_penalty, the
percent penalty each protein overdose gives.
Before, modules could only set the maximum number
of penalties, not the amount of each penalty,
which was set at 10%, when some modules need
other values.

The changes leave the default for all modules at
10%, so module creators have to modify the value
to change the game behavior.

Otherwise, the changes only add the value to
Module Editor and the Documentation.