Pet Auto Scaling
Change pet scaling whenever the number of pets
in the party changes, reloading the sprites, and
change the vertical positioning if needed.

Pets will scale anywhere from screen_width /
max_pets to screen_height / 2.

Change food to be half the width of the pets.

Scales pet movement with the pet size, normalized
to the original movement with MAX_PETS = 4 on
a 240x240 screen.

Removes 2 unused imports in constants, but adds
1 needed for rescaling on demand.